[WIP] Refactor EventsSheet to make it partially themable
