+++ newIDE/app/src/EventsSheet/EventsTree/ClassNames.js
@@ -6,5 +6,4 @@ export const largeSelectedArea = 'large-selected';
 
 export const background = 'background';
 
-export const container = 'gd-events-sheet';
 export const eventsTree = 'events-tree';
