+++ newIDE/app/src/EventsSheet/EventsTree/Renderers/GroupEvent.js
