+++ newIDE/app/src/EventsSheet/EventsTree/fold.png
