+++ newIDE/app/src/EventsSheet/EventsTree/style.css
@@ -33,7 +33,7 @@
 }
 
 .gd-events-sheet .rst__row {
-    /* The "landing pad" highlight box is done with positionL absolute.
+    /* The "landing pad" highlight box is done with position: absolute.
      * Ensure it will cover all the event row but not the scaffolding lines.
      */
     position: relative;
@@ -119,23 +119,10 @@
     border: 1px transparent solid;
 }
 
-.gd-events-sheet .selectable:hover {
-    background-color: rgba(0,0,0, 0.1);
-}
-
-.gd-events-sheet .selectable.selected {
-    background-color: rgba(0,0,0, 0.1);
-    border: 1px #4ab0e4 solid !important;
-}
-
 /**
  * Large selectable area (events)
  */
 .gd-events-sheet .large-selectable {
     box-sizing: border-box;
     border: 1px transparent solid;
-}
-
-.gd-events-sheet .large-selectable.large-selected {
-    border: 1px #4ab0e4 solid !important;
 }
