+++ newIDE/app/src/MainFrame/Editors/StartPage.spec.js
@@ -2,6 +2,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
+import defaultTheme from '../../UI/Theme/DefaultTheme';
 import StartPage from './StartPage';
 const gd = global.gd;
 
@@ -9,7 +10,7 @@ describe('StartPage', () => {
   it('renders the start page with no project opened', () => {
     const tree = renderer
       .create(
-        <MuiThemeProvider>
+        <MuiThemeProvider muiTheme={defaultTheme}>
           <StartPage project={null} />
         </MuiThemeProvider>
       )
@@ -21,7 +22,7 @@ describe('StartPage', () => {
     const project = gd.ProjectHelper.createNewGDJSProject();
     const tree = renderer
       .create(
-        <MuiThemeProvider>
+        <MuiThemeProvider muiTheme={defaultTheme}>
           <StartPage project={project} />
         </MuiThemeProvider>
       )
